Various updates to get rest timer and other functionality

+++ WorkoutTracker_v2/Model/Exercise.swift
@@ -13,7 +13,7 @@ struct Exercise: Decodable, Identifiable {
     var sets    : Int
     var minReps : Int
     var maxReps : Int
-    var weight  : Int
+    var weight  : Float
 }
 
 extension MockData {
